backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0

# Image Processing
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3

# AI/ML (lightweight for mock service)
torch==2.1.2
torchvision==0.16.2
scikit-learn==1.4.0

# Data & Validation
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Caching & Performance
redis==5.0.1
aioredis==2.0.1

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Utilities
python-magic==0.4.27
python-dateutil==2.8.2

# Monitoring & Logging
loguru==0.7.2
